requirements-txt.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
matplotlib>=3.5.0
seaborn>=0.12.0
numpy>=1.21.0
pyyaml>=6.0
tqdm>=4.65.0